switched to serializers validation & added HTTP status codes

+++ django_react_app/django_backend/api/views.py
@@ -1,18 +1,18 @@
 import json
 import uuid
 from django.contrib.auth import get_user_model
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 from rest_framework_simplejwt.views import TokenObtainPairView
-from PIL import Image
 from . models import UserEvents
 from . models import UserReviews
 from . serializers import UserEventsSerializer
 from . serializers import GetReviewsSerializer
 from . serializers import GetProfilePictureSerializer
-from . import forms
+from . import serializers
 import ticketmaster_api
 import openweathermap_api
 User = get_user_model()
@@ -38,7 +38,7 @@ def getRoutes(request):
     """
     routes = [
         {
-            'Endpoint': '/api/signup_user/',
+            'Endpoint': '/api/signup/',
             'Method': ['POST'],
             'Restricted': False,
             'Description': {'POST': 'Signup a user and save their information inside the database'}
@@ -82,7 +82,7 @@ def getRoutes(request):
             'Description': {'GET': 'Get the reviews left by all users for that specific event'}
         },
         {
-            'Endpoint': '/api/profile/',
+            'Endpoint': '/api/profile/events/',
             'Method': ['GET', 'POST', 'DELETE'],
             'Restricted': True,
             'Description': {'GET': 'Returns an array of events from which the user have save',
@@ -119,17 +119,17 @@ def getRoutes(request):
     return Response(routes)
 
 @api_view(["POST"])
-def signup_user(request):
+def signup(request):
     """
-    Endpoint: /api/signup_user/
+    Endpoint: /api/signup/
     """
     user_info = json.loads(request.body)
-    form = forms.Signup(user_info)
-    if form.is_valid():
-        user = User.objects.create_user(username=form.cleaned_data["username"], password=form.cleaned_data["password"], zip_code=form.cleaned_data["zip_code"])
-        return Response(True)
+    validate = serializers.SignUpValidateSerializer(data=user_info)
+    if validate.is_valid():
+        user = User.objects.create_user(username=validate.validated_data["username"], password=validate.validated_data["password"], zip_code=validate.validated_data["zip_code"])
+        return Response(status=status.HTTP_201_CREATED)
     else:
-        return Response(form.errors.values())
+        return Response(validate.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -139,7 +139,10 @@ def events(request, page):
     """
     user = request.user
     events = ticketmaster_api.getEvents(user.zip_code, page)
-    return Response(events)
+    if events != False:
+        return Response(events, status=status.HTTP_200_OK)
+    else:
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -148,7 +151,10 @@ def eventsSearchInput(request, input, page):
     Endpoint: /api/events/search/input/<str:input>/page/<str:page>/
     """
     events = ticketmaster_api.getEvents(input, page)
-    return Response(events)
+    if events != False:
+        return Response(events, status=status.HTTP_200_OK)
+    else:
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -157,7 +163,7 @@ def eventsDetails(request, id):
     Endpoint: /api/events/details/id/<str:id>/
     """
     events = ticketmaster_api.getEventsDetails(id)
-    return Response(events)
+    return Response(events, status=status.HTTP_200_OK)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -166,7 +172,7 @@ def eventsWeather(request, latitude, longitude):
     Endpoint: /api/events/weather/latitude/<str:latitude>/longitude/<str:longitude>/
     """
     events = openweathermap_api.getEventsWeather(latitude, longitude)
-    return Response(events)
+    return Response(events, status=status.HTTP_200_OK)
 
 @api_view(["POST", "PUT", "DELETE"])
 @permission_classes([IsAuthenticated])
@@ -179,7 +185,7 @@ def userReviews(request):
         data = json.loads(request.body)
         review = UserReviews(event_id=data["event_id"], title=data["title"], userRating=data["userRating"], userComment=data["userComment"], user=user)
         review.save()
-        return Response(True)
+        return Response(status=status.HTTP_201_CREATED)
     
     if request.method == "PUT":
         data = json.loads(request.body)
@@ -187,13 +193,13 @@ def userReviews(request):
         review.userComment = data["userComment"]
         review.userRating = data["userRating"]
         review.save()
-        return Response(True)
+        return Response(status=status.HTTP_200_OK)
 
     if request.method == "DELETE":
         data = json.loads(request.body)
         review = UserReviews.objects.all().filter(id=data["review_id"])
         review.delete()
-        return Response(True)
+        return Response(status=status.HTTP_200_OK)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -203,25 +209,25 @@ def getAllReviews(request, event_id):
     """
     reviews = UserReviews.objects.all().filter(event_id=event_id).order_by("id")
     serializer = GetReviewsSerializer(reviews, many=True)
-    if serializer.data == []:
-        return Response(False)
+    if serializer.data != []:
+        return Response(serializer.data, status=status.HTTP_200_OK)
     else:
-        return Response(serializer.data)
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
 @api_view(["GET", "POST", "DELETE"])
 @permission_classes([IsAuthenticated])
-def profile(request):
+def profileEvents(request):
     """
-    Endpoint: /api/profile/
+    Endpoint: /api/profile/events/
     """
     if request.method == "GET":
         user = request.user
         data = UserEvents.objects.all().filter(user=user.id)
         serializer = UserEventsSerializer(data, many=True)
-        if serializer.data == []:
-            return Response(False)
+        if serializer.data != []:
+            return Response(serializer.data, status=status.HTTP_200_OK)
         else:
-            return Response(serializer.data)
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == "POST":
         user = User.objects.get(id=request.user.id)
@@ -229,9 +235,9 @@ def profile(request):
         if UserEvents.objects.all().filter(user=user.id, event_id=data["event_id"]).exists() == False:
             event = UserEvents(event_id=data["event_id"], title=data["title"], date=data["date"], city=data["city"], imageUrl=data["imageUrl"], minPrice=data["minPrice"], maxPrice=data["maxPrice"], user=user)
             event.save()
-            return Response(True)
+            return Response(status=status.HTTP_201_CREATED)
         else:
-            return Response(False)
+            return Response(status=status.HTTP_409_CONFLICT)
     
     if request.method == "DELETE":
         user = request.user
@@ -239,9 +245,9 @@ def profile(request):
         if UserEvents.objects.all().filter(user=user.id, event_id=data["event_id"]).exists():
             event = UserEvents.objects.get(user=user.id, event_id=data["event_id"])
             event.delete()
-            return Response(True)
+            return Response(status=status.HTTP_200_OK)
         else:
-            return Response(False)
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -252,10 +258,10 @@ def profileReview(request):
     user = request.user
     reviews = UserReviews.objects.all().filter(user=user.id).order_by("id")
     serializer = GetReviewsSerializer(reviews, many=True)
-    if serializer.data == []:
-        return Response(False)
+    if serializer.data != []:
+        return Response(serializer.data, status=status.HTTP_200_OK)
     else:
-        return Response(serializer.data)
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
 @api_view(["GET", "PUT"])
 @permission_classes([IsAuthenticated])
@@ -266,20 +272,22 @@ def profilePicture(request):
     if request.method == "GET":
         user = User.objects.get(username=request.user)
         serializer = GetProfilePictureSerializer(user, many=False)
-        if serializer.data == []:
-            return Response(False)
+        if serializer.data["profile_picture"] != None:
+            return Response(serializer.data, status=status.HTTP_200_OK)
         else:
-            return Response(serializer.data)
+            return Response(status=status.HTTP_404_NOT_FOUND)
         
     if request.method == "PUT":
-        file = request.data["file"]
-        file.name = str(request.user) + "_id_" + str(uuid.uuid4())[:8] + "." + file.name.split(".")[-1]
-        try:
-            Image.open(file)
+        file = request.FILES
+        validate = serializers.CheckFileValidateSerializer(data=file)
+        if validate.is_valid():
+            validate.validated_data["file"].name = str(request.user) + "_id_" + str(uuid.uuid4())[:8] + "." + validate.validated_data["file"].name.split(".")[-1]
             user = User.objects.get(username=request.user)
             user.profile_picture.delete()
-            user.profile_picture = file
+            user.profile_picture = validate.validated_data["file"]
             user.save()
-            return Response(True)
-        except:
-            return Response(False)
+            return Response(status=status.HTTP_200_OK)
+        else:
+            # status code 415 (UNSUPPORTED_MEDIA_TYPE) and or 413 (REQUEST_ENTITY_TOO_LARGE) would work
+            # my problem is that it is one of the two OR both. i'll choose 400 for now. i'll find a better solution later.
+            return Response(validate.errors, status=status.HTTP_400_BAD_REQUEST)
